# Enable use of background threads in Splinter. Engage tests to verify basic stability of this existing feature.