--- a/tests/functional/btree_test.c
+++ b/tests/functional/btree_test.c
@@ -1471,7 +1471,7 @@ btree_test(int argc, char *argv[])
    bool                   run_perf_test;
    platform_status        rc;
    uint64                 seed;
-   task_system           *ts;
+   task_system           *ts = NULL;
    test_message_generator gen;
 
    if (argc > 1 && strncmp(argv[1], "--perf", sizeof("--perf")) == 0) {
@@ -1503,6 +1503,8 @@ btree_test(int argc, char *argv[])
    rc = platform_heap_create(platform_get_module_id(), 1 * GiB, &hh, &hid);
    platform_assert_status_ok(rc);
 
+   uint64 num_bg_threads[NUM_TASK_TYPES] = {0}; // no bg threads
+
    data_config  *data_cfg;
    trunk_config *cfg = TYPED_MALLOC(hid, cfg);
 
@@ -1514,6 +1516,8 @@ btree_test(int argc, char *argv[])
                         &log_cfg,
                         &seed,
                         &gen,
+                        &num_bg_threads[TASK_TYPE_MEMTABLE],
+                        &num_bg_threads[TASK_TYPE_NORMAL],
                         config_argc,
                         config_argv);
 
@@ -1556,10 +1560,7 @@ btree_test(int argc, char *argv[])
       goto free_iohandle;
    }
 
-   uint8 num_bg_threads[NUM_TASK_TYPES] = {0}; // no bg threads
-
-   rc = test_init_task_system(
-      hid, io, &ts, cfg->use_stats, FALSE, num_bg_threads);
+   rc = test_init_task_system(hid, io, &ts, cfg->use_stats, num_bg_threads);
    if (!SUCCESS(rc)) {
       platform_error_log("Failed to init splinter state: %s\n",
                          platform_status_to_string(rc));