--- a/tests/functional/filter_test.c
+++ b/tests/functional/filter_test.c
@@ -318,6 +318,9 @@ filter_test(int argc, char *argv[])
    rc = platform_heap_create(platform_get_module_id(), 1 * GiB, &hh, &hid);
    platform_assert_status_ok(rc);
 
+   uint64 num_memtable_bg_threads_unused = 0;
+   uint64 num_normal_bg_threads_unused   = 0;
+
    trunk_config *cfg = TYPED_MALLOC(hid, cfg);
 
    rc = test_parse_args(cfg,
@@ -328,6 +331,8 @@ filter_test(int argc, char *argv[])
                         &log_cfg,
                         &seed,
                         &gen,
+                        &num_memtable_bg_threads_unused,
+                        &num_normal_bg_threads_unused,
                         config_argc,
                         config_argv);
    if (!SUCCESS(rc)) {