This commit exercises use of background threads, a feature that has been implemented but hadn't been fully tested & engaged.

- Minor adjustments are done to task system initialization to manage the order in which background threads are started up.

- Fixed some checks in task_system_test to cope with the new variation that this test case could, optionally, be exercised with `--num-normal-bg-threads` | `--num-memtable-bg-threads` arguments.

- Enhanced config_parse() to recognize these two new options.
- Added a case to test.sh to exercise the task_system_test with background threads.

-  Stabilization fixes to get large inserts stress tests working. (New test withdrawn)

-  Add new test cases to test.sh, exercising background threads.

---------------
**NOTE**: Before you can run the `large_inserts_stress_test` that was leveraged to exercise background threads functionality,  you need to pull-in in-flight fix for issue https://github.com/vmware/splinterdb/issues/458; Otherwise you will run into that assertion first. 

The dev-branch were this repro has been constructed, `agurajada/478-Enable-bg-thread-feature`, pulls-in that commit already.