@gapisback 2022-11-07
NOTE: Enabling background threads, if it does not run into the issues reported earlier, this feature does seem to show substantial gains in insert performance. 

For instance, with 4 inserting threads, and `--num-bg-threads 2` and `--num-memtable-bg-threads 1`, these two cases show:

- test_seq_key_seq_values_inserts_threaded: 25% improved throughput averaged across 4 inserting threads
- test_seq_key_seq_values_inserts_threaded_same_start_keyid: 35% improved throughput averaged across 4 inserting threads

Here is the evidence:

**Without background threads**: 

`$ build/release/bin/unit/large_inserts_bugs_stress_test --num-inserts 20000000 --num-threads 4 test_seq_key_seq_values_inserts_threaded`

```
ip-172-31-29-15:[48] $ build/release/bin/unit/large_inserts_bugs_stress_test --num-inserts 20000000 --num-threads 4 test_seq_key_seq_values_inserts_threaded
Running 1 CTests, suite name 'large_inserts_bugs_stress', test case 'test_seq_key_seq_values_inserts_threaded'.
TEST 1/2 large_inserts_bugs_stress:test_seq_key_seq_values_inserts_threaded Fingerprint size 29 too large, max value size is 5, setting to 27
fingerprint_size: 27
filter-index-size: 256 is too small, setting to 512
exec_worker_thread()::495:Thread 1  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141578, Thread-ID=1, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 3  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 40000000 (40 million) ...
exec_worker_thread()::495:Thread 4  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 60000000 (60 million) ...
OS-pid=141578, Thread-ID=4, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 2  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 20000000 (20 million) ...
OS-pid=141578, Thread-ID=2, Insert small-width sequential values of different lengths.
OS-pid=141578, Thread-ID=3, Insert small-width sequential values of different lengths.
exec_worker_thread()::601:Thread-3 Inserted 20 million KV-pairs in 25 s, 800000 rows/s
exec_worker_thread()::601:Thread-1 Inserted 20 million KV-pairs in 27 s, 740740 rows/s
exec_worker_thread()::601:Thread-2 Inserted 20 million KV-pairs in 27 s, 740740 rows/s
exec_worker_thread()::601:Thread-4 Inserted 20 million KV-pairs in 27 s, 740740 rows/s
Allocated at unmount: 3948 MiB
[OK]
TEST 2/2 large_inserts_bugs_stress:test_seq_key_seq_values_inserts_threaded_same_start_keyid Fingerprint size 29 too large, max value size is 5, setting to 27
fingerprint_size: 27
filter-index-size: 256 is too small, setting to 512
exec_worker_thread()::495:Thread 1  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141578, Thread-ID=1, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 2  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141578, Thread-ID=2, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 3  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141578, Thread-ID=3, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 4  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141578, Thread-ID=4, Insert small-width sequential values of different lengths.
exec_worker_thread()::601:Thread-4 Inserted 20 million KV-pairs in 21 s, 952380 rows/s
exec_worker_thread()::601:Thread-3 Inserted 20 million KV-pairs in 22 s, 909090 rows/s
exec_worker_thread()::601:Thread-1 Inserted 20 million KV-pairs in 22 s, 909090 rows/s
exec_worker_thread()::601:Thread-2 Inserted 20 million KV-pairs in 25 s, 800000 rows/s
Allocated at unmount: 1451 MiB
[OK]
RESULTS: 2 tests (2 ok, 0 failed, 0 skipped) ran in 53271 ms
```

**With background threads enabled**: 

`$ build/release/bin/unit/large_inserts_bugs_stress_test --num-inserts 20000000 --num-threads 4 --num-bg-threads 2 --num-memtable-bg-threads 1 test_seq_key_seq_values_inserts_threaded`

```
ip-172-31-29-15:[61] $ build/release/bin/unit/large_inserts_bugs_stress_test --num-inserts 20000000 --num-threads 4 --num-bg-threads 2 --num-memtable-bg-threads 1 test_seq_key_seq_values_inserts_threaded
Running 1 CTests, suite name 'large_inserts_bugs_stress', test case 'test_seq_key_seq_values_inserts_threaded'.
TEST 1/2 large_inserts_bugs_stress:test_seq_key_seq_values_inserts_threaded Fingerprint size 29 too large, max value size is 5, setting to 27
fingerprint_size: 27
filter-index-size: 256 is too small, setting to 512
Splinter task system created 1 background threads of type 'TASK_TYPE_MEMTABLE'.
Splinter task system created 2 background threads of type 'TASK_TYPE_NORMAL'.
exec_worker_thread()::495:Thread 4  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141859, Thread-ID=4, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 5  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 20000000 (20 million) ...
OS-pid=141859, Thread-ID=5, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 6  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 40000000 (40 million) ...
OS-pid=141859, Thread-ID=6, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 7  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 60000000 (60 million) ...
OS-pid=141859, Thread-ID=7, Insert small-width sequential values of different lengths.
exec_worker_thread()::601:Thread-7 Inserted 20 million KV-pairs in 21 s, 952380 rows/s
exec_worker_thread()::601:Thread-5 Inserted 20 million KV-pairs in 21 s, 952380 rows/s
exec_worker_thread()::601:Thread-4 Inserted 20 million KV-pairs in 21 s, 952380 rows/s
exec_worker_thread()::601:Thread-6 Inserted 20 million KV-pairs in 21 s, 952380 rows/s
Allocated at unmount: 3954 MiB
[OK]
TEST 2/2 large_inserts_bugs_stress:test_seq_key_seq_values_inserts_threaded_same_start_keyid Fingerprint size 29 too large, max value size is 5, setting to 27
fingerprint_size: 27
filter-index-size: 256 is too small, setting to 512
Splinter task system created 1 background threads of type 'TASK_TYPE_MEMTABLE'.
Splinter task system created 2 background threads of type 'TASK_TYPE_NORMAL'.
exec_worker_thread()::495:Thread 4  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141859, Thread-ID=4, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 5  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141859, Thread-ID=5, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 6  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141859, Thread-ID=6, Insert small-width sequential values of different lengths.
exec_worker_thread()::495:Thread 7  inserts 20000000 (20 million), sequential key, seqential value, KV-pairs starting from 0 (0) ...
OS-pid=141859, Thread-ID=7, Insert small-width sequential values of different lengths.
exec_worker_thread()::601:Thread-4 Inserted 20 million KV-pairs in 16 s, 1250000 rows/s
exec_worker_thread()::601:Thread-7 Inserted 20 million KV-pairs in 16 s, 1250000 rows/s
exec_worker_thread()::601:Thread-5 Inserted 20 million KV-pairs in 17 s, 1176470 rows/s
exec_worker_thread()::601:Thread-6 Inserted 20 million KV-pairs in 17 s, 1176470 rows/s
Allocated at unmount: 1084 MiB
[OK]
RESULTS: 2 tests (2 ok, 0 failed, 0 skipped) ran in 39955 ms
```